[#3]/refactor TutorialContents 관심사 분리
